fix typos, rename error class
