+++ class/ReactNativeBlobUtilCanceledFetchError.js
@@ -1,10 +1,10 @@
+const CANCELED_FETCH_ERROR_NAME = 'ReactNativeBlobUtilCanceledFetch'
+
 class CanceledFetchError extends Error {
     constructor(message) {
         super(message);
-        this.name = CanceledFetchErrorName;
+        this.name = CANCELED_FETCH_ERROR_NAME
     }
 }
-
-export const CanceledFetchErrorName = 'ReactNativeBlobUtilCanceledFetch'
 
 export default CanceledFetchError
