+++ fetch.js
@@ -4,7 +4,7 @@ import fs from "./fs";
 import getUUID from "./utils/uuid";
 import { DeviceEventEmitter } from "react-native";
 import { FetchBlobResponse } from "./class/ReactNativeBlobUtilBlobResponse";
-import { CanceledFetchError } from ".class/ReactNativeBlobUtilCanceledFetchError";
+import CanceledFetchError from "./class/ReactNativeBlobUtilCanceledFetchError";
 import ReactNativeBlobUtil from "./codegenSpecs/NativeBlobUtils";
 
 const emitter = DeviceEventEmitter;
