+++ index.js
@@ -15,7 +15,7 @@ import ios from './ios';
 import JSONStream from './json-stream';
 import {config, fetch} from './fetch';
 import URIUtil from "./utils/uri";
-import { CanceledFetchError } from ".class/ReactNativeBlobUtilCanceledFetchError";
+import CanceledFetchError from "./class/ReactNativeBlobUtilCanceledFetchError";
 
 const {
     ReactNativeBlobUtilSession,
